Android-Camera: Handle IllegalStateException in setRepeatingRequest
[0]setRepeatingRequest is used to continuously stream frames. This
method may throw an IllegalStateException if the session is no longer
active—either because it was explicitly closed, a new session was
created, or the camera device was closed.

While IllegalStateException is already handled in other calls to
setRepeatingRequest, it was missing in the
StillPhotoCaptureSessionCallback.

This commit adds proper handling for IllegalStateException in that
callback. Additionally, it introduces a check to ensure mIsStarted is
set before invoking setRepeatingRequest.

[0]https://developer.android.com/reference/android/hardware/camera2/CameraCaptureSession#setRepeatingRequest(android.hardware.camera2.CaptureRequest,%20android.hardware.camera2.CameraCaptureSession.CaptureCallback,%20android.os.Handler)

Pick-to: 6.9 6.8
Fixes: QTBUG-136227
Change-Id: I733d81761f94aea6da267bae6d339727c39a858c
Reviewed-by: Nils Petter Skålerud <nils.petter.skalerud@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>