Android: Use QCachedValue in OpenSLESEngine class
This commit is part of OpenSLESEngine class refactoring.

QCachedValue class is used for thread-safe access and lazy
initialization. That is exactly what is needed for supported audio
settings getters in OpenSLESEngine class.

This change simplifies the implementation and ensures safe concurrent
access to supported audio input/output configuration.

Pick-to: 6.9 6.8
Task-number: QTBUG-136781
Change-Id: I943b391a8b800c61d9c7bd5cd08d11361dd0c546
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>