CMake: Add PURL and CPE info to 3rd party attribution files
See commit description of 335ee775db270158455d9fbf89db5c7671003d8a
why there are no PURLs and CPEs for libjpeg, zlib and boost that come
as part of ffmpeg.

For ffmpeg, we remove the prefix 'n' from the version, and keep just
the version number, because then it's easier to reuse the value in
both the PURL and CPE fields, which require different values.

[ChangeLog][Third-Party Code] Added PURL and CPE information to the
attribution files of 3rd party sources.

Pick-to: 6.5 6.8 6.9 6.10
Fixes: QTBUG-137259
Task-number: QTBUG-129602
Change-Id: I1b9ab20283eb68a6103be6b51006ff2d9f77d8bf
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>