CMake: remove wmsdk detection
wmsdk.h is not used from within QtMultimedia and it is a legacy API
anyway. So the compile-time check is not needed. Furthermore, the audio
backend is available on all windows platforms, so we can update the
detection as well.

Pick-to: 6.9 6.8
Change-Id: I2e04ea4874fbc70c9d766feadd916046d55dc3f1
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 24ba2b298c153553ef1ae8eec69c699e6c0058e9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>