QAudioSource, Windows: Call requestStop() when error occurs
When disconnecting a device during a capture session,
IAudioCaptureClient::GetBuffer fails with the error code
AUDCLNT_E_DEVICE_INVALIDATED. This causes us to stop the audio-client,
but we will hit an assert within ~QPlatformAudioIOStream().

This patch makes sure we all requestStop(), to signal that we have
cleaned up the resources.

Pick-to: 6.10 6.9 6.8
Task-number: QTBUG-137150
Change-Id: If6a85783ff1375b19de61323b0cc3e901b67e45d
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>