Android, Camera: Remove unused capture request template parameter
In Android Camera2, we can use a template as a baseline for our preview
capture request. In the code it's been hardcoded to a magic value "3",
which maps to the value "CameraDevice.TEMPLATE_RECORD" in Android SDK.

This parameter is not used anywhere and makes for less readability.
This patch removes this parameter.

No functional changes.

Pick-to: 6.9 6.8
Task-number: QTBUG-130636
Change-Id: Ie225a26f05c72feb2b17178a3a6581ffbab6d4c0
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 09639c7b7fcf425801748bffa7137779d94570fe)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>