Android: Allow QImageCapture to recover from camera still photo failing
Currently, if a camera still photo capture fails, the QImageCapture
object is left waiting for an image indefinitely. Additionally, it
currently does not support queuing new image requests. This means
that whenever a still photo fails, it becomes impossible to take new
images using the QImageCapture object.

This patch does the following:
- Introduces internal methods to QFFmpegImageCapture to allow removing
a pending image from the queue
- Introduces a new QAndroidCamera event to report a failed still photo
image capture
- Connects the two

This allows a QImageCapture to stay in working state even when still
photo capture fails.

Pick-to: 6.10 6.9 6.8
Change-Id: I19b9b2d9e24a376f1372ec061b3a6213d1c58adc
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>