Android, camera: Unconditionally apply zoom
In the current code, we only apply the zoom factor to the Android
camera session if the QCamera setting is not 1.0. This is error-prone
because the camera is allowed to keep the zoom factor applied
by other commands, such as taking a still photo.

This patch makes us unconditionally apply the zoom factor, so we can
trust that it's always going to be applied correctly.

Pick-to: 6.9 6.8
Change-Id: I2b1f4d9b8d657469a585e958a33ec3e7c00b8552
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit c248abc28e4bc3130c563f35a706f13a32083411)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>