Android, camera: Create Java-side CameraSettings helper class
The Java class QtCamera2 holds variables that control what settings
to apply to the camera. These variables can be changed at any time from
the QCamera thread. At the same time, still photo capture is a routine
that runs over time in order to finalize settings such as auto-focus
and auto-exposure calibration, which will read these values.

This creates an issue where commands such as QCamera::setFocusMode can
break the still-photo capturing by overriding the on-going
capture-request, or update the focus-mode while the still-photo routine
is reading from it.

This patch lays the groundwork for a solution where we can easily take
a snapshot of all camera-settings at a point in time, in a thread safe
manner and pass it around where necessary. It introcudes a helper class
'CameraSettings' which holds settings we need to apply to the camera,
and a method that allows us to create a deep-copy of the currently
stored CameraSettings object so that we can pass it around.

We can then build on this to use it more correctly during still photo
capture. See relation chain for context.

Pick-to: 6.9 6.8
Task-number: QTBUG-130636
Change-Id: If6305857fec70e457c5c20542a01ca565612634f
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 7590c2501b33f547a93a30e71db27335cd8fcc00)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>