Android: Move logic for applying preview settings to helper function
Setting up a CaptureRequest.Builder to become configured for previewing
is currently hardcoded into the ".start()" method.

This patch moves this logic into a helper function for improved
readability, and allows us to reuse to the code in the future. See
relation chain for context.

No functional changes.

Pick-to: 6.10 6.9 6.8
Task-number: QTBUG-130636
Change-Id: Ia7defb2ffc57d1087fdf5c5fa1cdfde69f815f97
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>