Android, camera: Consolidate logic for going into preview mode
There are currently two events in which we have to enter regular
preview mode for the camera: when the camera goes into active mode,
and when we finish a still photo.

This patch moves the logic into a dedicated function and reuses the
function in both events.

Pick-to: 6.9 6.8
Change-Id: Idd370ce68a6789073c08d9b651bc948da53f8a49
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 3e603ae9b54ee0661fba3f54c077b96d48b822a4)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>