Android, camera: Move precapture waiting logic into dedicated class
In order to take a good still photo, we want to trigger auto-focus
and auto-exposure calibration and wait for them to lock in before we
capture the image. This waiting logic is currently happening inside
the same request-callback class as the one that is used by regular
previewing.

The error-handling logic and the waiting logic when doing still photo
is very different from the preview logic. This patch moves the still
photo behavior into a dedicated class. This makes it easier for us
to add more logic to the still photo routine in the future. See
relation chain for context.

Pick-to: 6.10 6.9 6.8
Task-number: QTBUG-130636
Change-Id: I90376213fa74d4865b7aa9e8d4f083ce2b863770
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>