Android, camera: Allow precapture to take a copy of CameraSettings
This patch allows the still photo precapture process to take a copy
of all the CameraSettings that were in place when the still photo
process begun. It can then store this and pass it on when we finalize
the photo.

This allows us to avoid an issue where the CameraSettings change values
between the still photo process starting, and it ending. This can
i.e happen if QCamera::setFlash is called on the C++ thread.

Pick-to: 6.9 6.8
Task-number: QTBUG-130636
Change-Id: I0991f62d77f3df5d555565baf18e3dd264dcb088
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit c373d6f605bcb14cabe7e23081f2eff4fc1d3646)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>