Fix 5 sec freeze in QGstreamerMediaPlayer destructor
When a QML VideoOutput is destroyed, its QGstreamerVideoSink is
replaced with a GstFakeSink. If the pipeline is in paused state, the new
fakesink cannot transition completely to PAUSED, probably because it
is waiting to receive a buffer in order to preroll.
If the QGstreamerMediaPlayer destructor is called at this point, its
call to setStateSync(GST_STATE_NULL) blocks until it times out, probably
because it is waiting for the stalled fakesink transition.

Seting async to false on the fakesink disables the need to wait for a
buffer in order to transition to PAUSED state, and thus prevents the
freeze.

Fixes: QTBUG-137875
Change-Id: I471c56339cbf85b4a771cc7bc2f3cbbc6f9cb5e1
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>