Android, camera: Simplify isAfModeSupported
The affected method used to be called from native code, and had the
behavior that it should return true if the camera supported a specific
auto-focus mode, and also had a working implementation for it.

This method is no longer being used from the C++ thread, and we can
simplify the rules to only consider if it's available, not whether
it has a working implementation.

Pick-to: 6.10 6.9 6.8
Change-Id: Ieed0c56cebfea0c32227b8927f978e1af6a41386
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>