Android, camera: Refactor still photo capture logic
This patch lays down the initial work for refactoring the still photo
capturing process in the Android FFmpeg camera code. It replaces
the existing unreliable logic that is meant to wait for auto-focus
and auto-exposure to settle.

The patch does the following:
- Lays the groundwork for a more flexible solution where we might
  be using auto-focus, auto-exposure, both or none of them.
- Adds commenting and classes that improve clarity on what is
  happening in what order during the still photo capturing routine.
- Consolidates the logic for configuring the still photo requests
  into a single helper function.

This patch has been tested to allow QCamera::FlashModeOn to work
correctly on Motorola, Samsung and Pixel devices.

Known issues: QCamera::FlashModeAuto still does not work. Will be
solved in upcoming patch. See relation chain.

Pick-to: 6.10 6.9 6.8
Task-number: QTBUG-130636
Change-Id: Ia82710455f000f3e387b1d45302625973859e79d
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>