Android, camera: Track when the camera is currently taking still photo
Currently when updating a camera setting, such as QCamera::setFocusMode,
will cause the new setting to be applied immediately to the
camera capture request. If there is an on-going still photo capture
in progress, this will break the still photo and it will never
complete.

This patch introduces a thread-safe bool that tells us whether we are
currently in the process of taking a still photo. We use this boolean
to tell us whether we should apply camera settings immediately.

Pick-to: 6.9 6.8
Task-number: QTBUG-130636
Change-Id: I76fd28b2161d8a949c15887632a71ddc030c6846
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
(cherry picked from commit 8614fce90a3cec80a7487e83e75eb649cfdca249)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>