Examples: Fix missing file extension in AudioRecorder on Android
The AudioRecorder example lets users set the output file name with its
extension. On Android, file name and extension are filled wit some
default values. This worked correctly for all formats except when
"Default file format" was selected - no extension was added in that
case.

This patch ensures that a proper file extension is used for the
"Default file format" option on Android.

Pick-to: 6.10 6.9 6.8
Fixes: QTBUG-98145
Change-Id: Ibccc3480808c2e9bdf31573bc31b6e1563724af5
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>