WASAPI: don't expose full sampling rate range
WASAPI only supports resampling to in the IAudioRenderClient, not in the
IAudioCaptureClient. So we only expose the rate that the device actually
supports.
We may need to revisit this in future to be somewhat consistent across
platforms

Fixes: QTBUG-138059
Pick-to: 6.9 6.8
Change-Id: I6a8814b37dc57830487aa52c0df7d3de514eea67
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>
(cherry picked from commit 2eef24a1da755022be1c5e0a9492b480e5e60eed)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>