Android, camera: Implement QCamera::FlashModeAuto
In the current code, Android reports QCamera::FlashModeAuto as
supported, but it behaves the same as QCamera::FlashModeOff.

This patch introduces initial support for this automatic flash
behavior, by adding logic that allows the camera to determine whether
the scene requires flash for proper lighting. If the scene was
determined to require flash, we resubmit the still photo capture
request as if it was using QCamera::FlashModeOn.

Tested to to work correctly on Pixel 8a and Samsung Galaxy A04.

Known issue: Under some circumstances, the exposure will be too bright
on Motorola Razr 40 Ultra.

Pick-to: 6.10 6.9 6.8
Fixes: QTBUG-130636
Change-Id: Ie56890032a439b299484785fbb5b2d5be7876a8a
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>