WASAPI: use windows resampler for channel layout conversions
Users expect to use open the QAudioSink for mono sources, even if the
device supports multiple channels.

Likewise QAudioSource is expected to support all sampling rates, while
WASAPI only supports the sampling rate of the backend.

We now support this in a non-realtime safe way by injecting the windows
resampler.

Pick-to: 6.8 6.9 6.10
Change-Id: I1db9942b1a423064d37d2a43f9242d5ac6e78771
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>