Keep previous QVideoFrameTextures alive until next frame is rendered
In cases where rhi.currentFrameSlot always returns 1, our call to
QVideoTextureHelper::createTextures deletes the previous textures, which
frees the associated memory buffer in case of dma-buf rendering. On imx6
this caused tearing artifacts because the decoder started writing a new
frame to the same buffer, which could still be read from at this point.

Fixed by moving the old textures to a member variable which is reset
when the new textures have been rendered.

Pick-to: 6.10 6.9 6.8
Change-Id: I16c757b8b201eb60dab7c51a83948cf661eab5a7
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>