Apple, camera: Move camera rotation tracking into utility class
This patch is part of a relation chain to improve thread safety in the
Apple camera backend. Camera rotation will be used in multiple classes
in the future, such as during still photo capture. See relation chain
for context.

This patch moves the rotation tracking into a dedicated utility class,
with proper RAII semantics.

Pick-to: 6.10 6.9 6.8
Change-Id: I2879131cadb7cb3ad5c2a247ce80768a9a60e0ec
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>