CMake: remove find_package for windows media foundations
Windows media foundations is part of the windows sdk, just like all
other windows system libraries. Looking for it explicitly is not
necessary and we link the required sublibraries into the final targets,
anyway.
So we remove the find_package, enable FEATURE_wmf by default and link
the required sublibraries explicitly.

Pick-to: 6.8 6.9 6.10
Change-Id: Iaa24e482b7a68340e8a0db4f19f270f492332187
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>