Windows: clean up handling of predefined symbols
INITGUID causes GUIDs to be emitted into the translation unit. We don't
want this define to leak into other translation units and we need to
prevent that it is defined after certain headers are include, which is a
struggle for unity builds. We therefore move the INITGUID define to the
build system, which will gracefully handle unity builds.
We furthermore don't redefine WINVER, but include qt_windows.h early and
statically assert the minimum requirement provided by that header.

Pick-to: 6.8 6.9 6.10
Change-Id: I0334eff853cd29729558d90a5184238efe3c8da1
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>