Doc: Fix incorrect usage of \generatelist
The \generatelist command takes an optional argument
that affect how the members are listed, for example,
`classesbymodule` which creates an annotated list of
the classes in a named C++ module.

`groupsbymodule` is not recognized or documented. QDoc
falls back to generating a list of all members of a
group in this case. While the end result is correct,
this relies on an undocumented feature that is likely
to change and break.

Replace these instances with an explicit \annotatedlist
command that produces the same output.

Task-number: QTBUG-138901
Change-Id: I26052bfc7ad9db2aad3204eb4080f16012161ebe
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit 4ac1e9309774b656b04db025ae66633d5d8e5773)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>