Android: Refactor JNI calls in audio devices
This patch amends an earlier commit that removed OpenSLES as a
dependency.

JNI calls in QAndroidAudioDevices are refactored to make use of the JNI
type declaration macros.

The preferred audio format for each device is now passed to
QAndroidAudioDevice as an argument instead of done in the constructor.

Android audio devices now have an isBluetoothDevice boolean that is set
on creation. This allows removing cached device arrays and obsolete
methods from QtAudioDeviceManager.

Task-number: QTBUG-137813
Pick-to: 6.10
Change-Id: I67da2375d0c8a0ccab95f8eb0c0b08b11b801668
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>