Docs, FFmpeg on Android: Add diffutils package
When building FFmpeg for Android, using Windows host, the FFmpeg
configure requires the tool 'cmp', which is not installed by default.
This tool is supplied by the diffutils package.

This patch adds this package to the list of required package,
for the documentation of building FFmpeg on Android.

Pick-to: 6.9 6.8
Change-Id: Ieab9c3c80b2fc7702c482a007057802ee91f9c60
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit deed61be38520a7e171a1ae051d366ae891e8550)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>