Android: Remove condition for AAudio low latency performance mode
We don't need to check for low latency audio support when building an
AAudio stream, we can just check for the relevant parameter after
opening the stream.

This patch also adds a categorized warning if low latency performance
mode is not used, as this can cause issues due to Android audio routing.

Pick-to: 6.10
Change-Id: Id836dfba8d268c3960f4071c54c21762dda76786
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>