Apple, camera: Cleanup surface transform logic
This patch is part of a relation chain to improve thread safety
in the Apple camera backend. The manner in which we generate
the surface transform is a inconsistent. The transform for the surface
is handled in the surfaceTransform() method on the QAVFCamera object.
Meanwhile, the presentation mirroring is handled in the frameHandler
callback.

This setup can be problmatic in the future patches when we harden
thread safety. This patch does the following:

- Makes surfaceTransform() a free function, safe to call from any
thread.
- Makes surfaceTransform return a struct that also contains the boolean
for whether mirroring should be applied during presentation, and
remove this logic from the frameHandler.

No functional changes.

Pick-to: 6.9 6.8
Change-Id: I90ba2036024d62368600d25117e4e4702d60d3cd
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 3a012c5db3af5acee6d050a0ade03c8b99410509)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>