GStreamer: Introduce QGstreamerRelayVideoSink
Changing video outputs of components that utilize an underlying
GStreamer pipeline was done by replacing the underlying video sink
element and reconfiguring the pipeline. For more complex pipelines, like
what one could achieve by using custom pipeline sources, this could
still often lead to broken pipelines which are difficult to debug. This
is why e.g. changing the video sink of a QMediaPlayer was disabled when
using a custom pipeline source.

This patch changes this logic so that QGstreamerMediaPlayer and
QGstreamerMediaCaptureSession create a fixed relay video sink as part of
the GStreamer pipelines which are not replaced, and are instead
connected to pluggable sinks which can be replaced on the fly.

This patch doesn't change anything for the audio sinks, which still keep
the limitation.

Fixes: QTBUG-134554
Pick-to: 6.10
Change-Id: Ie4c9fa4671ecbdaae2f5f078e6c6d27eb9b38311
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>