Apple: Pass AVCaptureConnection in sample-delegate's transform-provider
The QAVFSampleBufferDelegate object is used to receive frames from
AVFoundation. In addition, it holds a callback to generate a
frame-transform, to add information such as rotational metadata.
This is currently only used when receiving frames from a camera.

The current way this is done is problematic whenever we are trying
to read the rotation set by the AVCaptureConnection. Currently we
are grabbing AVCaptureConnection, stored by the
AVCaptureVideoDataOutput inside the QAVFCamera object. This can be
problematic since we are reading the QAVFCamera object from the
background thread, while making changes to the same object in the
foreground thread.

This patch makes the AVCaptureVideoDataOutput pass its own
AVCaptureConnection into the transform-provider callback,
so that we always make sure we are reading the right value.

Pick-to: 6.9 6.8
Change-Id: If9f05b48199cb680c96ec7d41556be74c1d28eec
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Mikko Hallamaa <mikko.hallamaa@qt.io>
(cherry picked from commit b504372dc1eb8a695ee8f751506f22e2ea4d8e3d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>