Apple, camera: Solve potential frameHandler race condition
In the current code, when receiving camera frames, we receive them
on a background thread through the QAVFSampleBufferDelegate. This
invokes a frame-handler callback which in turn captures a reference to
QAVFCamera or QAVFScreenCapture::Grabber, which live on the
foreground thread.

Particularly when stopping a capture session, or destroying the
QAVFCamera object, there is a small chance for there to still be
video-frames queued up on the background thread and referencing
the QAVFCamera object. There is little documentation on what
threading guarantees are offered here, so this is a potential race
condition.

This patch solves this by adding a method to QAVFSampleBufferDelegate
'discardFutureSamples' which allows us to clear the frame-handler
callback. The patch uses this method by providing a blocking
job to the background thread, so we try to flush as many frames as
possible before finishing resource cleanup.

Pick-to: 6.10 6.8
Change-Id: I9b6e31294ecd73dcc80ec541d6e9e27a303276a4
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>