Android: Prepare Android correctly when using a Bluetooth source
Using a Bluetooth input device for a QAudioSource wasn't working
properly, as it requires either starting Bluetooth SCO communication or
setting the Bluetooth device as a communication device (api 31+).

This patch modifies QtAudioDeviceManager::setAudioInput to be called
when we're about to create a source stream, and refactors it with
setAudioOutput to set a communication device or start SCO when needed.

Additionally, any preparations done when creating a stream are reverted
when destroying it.

Pick-to: 6.10 6.8
Fixes: QTBUG-126028
Change-Id: I1edbff98cdcea2903181723c43c41ed812b4b845
Reviewed-by: Bror Wetlesen Vedeld <bror.vedeld@qt.io>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>