QRtAudioEngine: Avoid running commands on finished voices
Finished voices are erased from the realtime voice registry in the audio
callback, so the assertion in runRtCommand methods that asserts the
voice is still present can trigger.

This patch replaces the assertion with a return when the voice is
removed.

Fixes: QTBUG-139993
Change-Id: I123bde67717e10dd0bea1b65b98c9944716f29c6
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 6d163487437aa093b4ac0434521a6851abae4151)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>