Pipewire: improve support for multiple audio formats
Some hdmi devices report multiple formats, i.e. "raw" and "iec938". Each
of the formats is enumerated independently, so we need to collect all
replies and add a sync barrier before we proceed with the next message.

If we receive a device with both "raw" and "iec938" formats, we will
prefer the "raw" format.

Pick-to: 6.10
Change-Id: I9f9c8a21cfd2cb0f848a8b534f3747e3c26d80f7
Reviewed-by: Bror Wetlesen Vedeld <bror.vedeld@qt.io>
Reviewed-by: Mikko Hallamaa <mikko.hallamaa@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>