AudioOutput: Decouple initializing audio from audio format application
previously we initialized and applied audio format in the same function

This means we could not apply an audio format as this requires us to
rebuild the generator and audiosink. Calling the function again would
just set the format to the default format, it would not allow us to
apply a new format.

This change decouples the initialization and the application of
audio format allowing us to send a custom format as a parameter.

Task-number: QTBUG-134263
Pick-to: 6.8
Change-Id: I4f64ab08e21938a657c843933011163b87eaffac
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit cdc489d2a39e209c1d681a2980ff12fe4d56da16)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>