Apple, camera: Make helper function 'findSuitableAVCaptureDeviceFormat'
The logic that applies camera format has a race condition, due to
writing to the QAVFSampleBufferDelegate while it is being read by the
background thread. Therefore, the method 'applyCameraFormat' needs to
be refactored to properly restart the necessary parts of the
AVCaptureSession.

As part of that refactor, this patch moves the logic that finds a
suitable AVCaptureDeviceFormat for a given QCamerFormat, into a helper
function.

Pick-to: 6.10 6.8
Change-Id: I48d129cd74059072eb6553d183c00a9fd8f46c02
Reviewed-by: Mikko Hallamaa <mikko.hallamaa@qt.io>
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>