Apple, camera: Make method for applying format to capture session
The logic that applies camera format has a race condition, due to
writing to the QAVFSampleBufferDelegate while it is being read by the
background thread. Therefore, the method 'applyCameraFormat' needs to
be refactored to properly restart the necessary parts of the
AVCaptureSession.

As part of that refactor, this patch moves the logic that applies
the format settings to the AVCaptureSession, into a dedicated helper
method.

Pick-to: 6.8
Change-Id: Ic0ca1ab855b695d6aec7e84aa5a0af5fdcfa8bb7
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
(cherry picked from commit 5b86a870783a76ccba6b6a5f088d47accb0769f1)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>