Android: Fix host audio buffer size when converting sample formats
When converting between sample formats, the chunk size for transferring
data between the ringbuffer and host buffer was calculated from our
ringbuffer which was set up with the sample format in QAudioFormat.
This caused an assertion failure in convertSampleFormat due to the
destination range being the same size as the source range.

This patch fixes the logic in ringbuffer consuming/production logic to
ensure the chunks to/from host buffer are correctly sized whether sample
format conversion is applied or not.

Additionally for Android where we utilize sample format conversion, we
need to take the native sample format into account when creating the
QSpan representing host buffer if sample format conversion is expected.

Fixes: QTBUG-140472
Pick-to: 6.10.0 6.10 6.8
Change-Id: I6c09ffd9156d6b08884cf3351a4d589386688996
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>