AudioOutput: Add feature for changing audio format
Previously we had no way to change the audioformat in this example

The audioformat is composed of three components: channel count,
format and sample rate. This adds three dropdowns that will allow a
user to set these independently, in case of an invalid format a
pop-up will occur and the gui will resync based on actual values.

Pick-to: 6.8 6.10
Fixes: QTBUG-134263
Change-Id: Ia0ee57ad347152aee3a840088849e810f58638af
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Nils Petter Skålerud <nils.petter.skalerud@qt.io>