Apple, camera: Initial cleanup of format logic
When applying formats to the capture session, we use the function
'updateCameraFormat'. This function currently has multiple points of
failure, few of which are handled. Additionally, it will write to
certain member variables before determining if the entire logic can
succeed.

This patch improves the error-handling of this function. It also
moves assignment of variables to the end, when we have determined that
the function has succeeded. It also removes the function
'setPixelFormat', as it also has dubious side-effects.

Pick-to: 6.8
Change-Id: I62f63c82e08aa1df1dc2517faeb6ba46141b5e95
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
(cherry picked from commit ae14361ef5f1fca6e2f796b3b556697665564d5e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>