Apple, camera: Split input-device into 'clear' and 'setup' methods
This patch is part of an on-going relation chain to improve behavior
and architecture of Apple camera code.

The current code uses a function 'refreshAvCaptureSessionInputDevice'
in order to clear and setup the camera against the AVCaptureSession.
This can be a bit hard to work with, and it's not clear when it will
set up or just tear down. This patch splits the methods into two
pieces, and now takes a parameter for AVCaptureDevice to setup.

Pick-to: 6.8
Change-Id: Icd60c334645aad0416e0f93988866325c8340a67
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
(cherry picked from commit bfeff37a2187144f1f85fb0db7d930e1150310ca)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>