Apple, camera: Improve AVCaptureVideoDataOutput thread safety
The current logic has AVCaptureVideoDataOutput reads rotational data by
accessing the QAVFCamera object. This can cause issues when swapping
devices, because the AVCaptureVideoDataOutput will have frames in
flight belonging to the previous camera device. It can also cause
issues whenever we are clearing the rotation-tracking from the
QAVFCamera object.

This patch fixes this by doing the following:

- Make a copy of the rotational tracker object and store it within the
AVCaptureVideoDataOutput object itself, so it has it's own thread-safe
copy.
- Create a new AVCaptureVideoDataOutput object whenever we start a
session, or switch devices. This means there is a chance having
two such objects alive for a short period, but they are scheduled
on the same background-thread.
- Solve race condition in format-switching logic

Pick-to: 6.10 6.8
Change-Id: Ia1b7311fec72cf8d843683d531f789d980cafaf2
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>