Apple, camera: Make helper for setting up and clearing configuration
There are several steps that must be done to start a new capture
session. Currently there is some code duplication to doing all these
steps.

This patch moves this logic into a single helper function with an
expected return value we can handle, this removes parts of the code
duplication.

As a side-effect, we now fix a bug caused by not setting the format
when calling QCamera.setActive(true), which can commonly occur if
QCamera.setCameraFormat was not called in between reconnecting the
physical device.

Pick-to: 6.10 6.8
Change-Id: Ia26b58bdb917f08d497528ada746f86c78016058
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>