Apple: Move QAVFCamera construction into factory function
It is useful for us to be able to use Objective-C syntax in several
of the class definitions that interact with AVFoundation. Currently,
the class header for QAVFCamera is included directly into the
file 'qffmegmediaintegration.cpp', which stops us from using the
Objective-C syntax.

This patch moves construction into a dedicated factory function, so
we don't need to include the QAVFCamera header directly in
'qffmpegmediaintegration.cpp'. See relation chain for context.

Pick-to: 6.10 6.8
Change-Id: I51c67e3843d5ca1368f73e455c0a1c5cdf3c2685
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>