Update tst_qmediaformatbackend to include WebM as supported format
FFmpeg n8.0 treats WebM as a supported format and associates it with the
AV1 video codec and the Opus/Vorbis audio codecs. Since these audio
codecs are considered experimental, they are not enabled by default in
our implementation. As a result, QMediaFormat may be constructed with an
unspecified audio codec. This case should be treated as valid, and the
tests have been updated accordingly.

Task-number: QTBUG-139714
Pick-to: 6.10
Change-Id: I4d5b24acfd7db7a5ba5082095df31538f44912b7
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>