AudioRecorder: Remove Audio Level GUI element and related code
Previously we had an unimplemented Audiolevel gui element in this app

This looks like it was intended to be implemented in 2017 but the
future implementation was scrapped. This patch removes all code
related to it from the .ui file and src files.

Fixes: QTBUG-140658
Pick-to: 6.8
Change-Id: I63fab7aaede6f58bfa7ed44d4f28efaefcb5e128
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit fcb411c17dee9fa809f164ac07f1cebdc3f68cb1)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>