Apple, camera: Fix rare crash when switching camera-device
It is not legal to call AVCaptureSession.stopRunning() when we are
currently performing a AVCaptureSession configuration using the method
AVCaptureSession.beginConfiguration(). Doing so results in a crash.

The recent changes to the Apple camera code introduced a rare crash
where this happened when switching devices. This can be
reproduced whenever the old and new camera-devices have no formats
in common.

This patch fixes this by moving the AVCaptureSession's configuration
logic inside a lambda, and then calls 'stopRunning()' if that lambda
returns a q23::unexpected.

Change-Id: Ifdcb1b687a319f16a4291b03aaa7d6c21ab1d4a0
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Bror Wetlesen Vedeld <bror.vedeld@qt.io>
(cherry picked from commit 82f596711865f850e49c1a235a7787e7c44ef443)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>