declarative-camera, docs: Use \snippet for code snippets
The example 'declarative-camera' has a documentation page associated
with it that includes code-snippets pulled directly from the examples
source code. The QDoc parts that pull these code-snippets, rely on a
error-prone skipto and printuntil logic, which tends to fall apart
when we move code around or change indentations. Modern QDoc usage is
to use the propert \snippet command.

This patch replaces all code-snippets with the \snippet command.

As a side-effect, this fixes indentation issues in the documentation
page.

Pick-to: 6.10 6.8
Change-Id: I7a43080667e08488ca5688bee185e6d20f3c69ec
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>